pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sage"
version = "0.1.0"
description = "SAGE - AI tutor that teaches through conversation, not curriculum"
readme = "docs/narrative/SAGE_narrative.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Chad Timothy"}
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "openai>=1.0",
    "typer>=0.9",
    "rich>=13.0",
    "websockets>=12.0",
    "python-jose[cryptography]>=3.3",
    "email-validator>=2.0",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
sage = "sage.api.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/sage"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/data",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "performance: marks tests as performance tests (may be slow)",
]
